+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 Version 0.0.6
 =============
 
+* Removed redundant payment method field on the Order model.
 * The "thank you" view does not crash anymore when it's refreshed. It now
   displays the last order the user placed.
 * Fixed a bug in the shippingbilling view where the returned form was a from